requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
joblib>=1.3
PyYAML>=6.0
shap>=0.44
matplotlib>=3.8